building/requirements.txt
Flask==1.0.2
google-cloud-language
